# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API para el sistema interno de Mariana Textil.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /proveedores/resumen:
    get:
      operationId: getProveedoresResumen
      tags: [proveedores]
      summary: Resumen agregado de todos los proveedores con saldos
      responses:
        "200":
          description: Resumen general de proveedores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProveedoresResumen"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /proveedores:
    get:
      operationId: listProveedores
      tags: [proveedores]
      summary: Lista proveedores con métricas de compras
      responses:
        "200":
          description: Objeto con resumen y lista de proveedores con métricas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProveedoresListResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createProveedor
      tags: [proveedores]
      summary: Crea un proveedor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProveedorInput"
      responses:
        "201":
          description: Proveedor creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proveedor"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /proveedores/{id}:
    get:
      operationId: getProveedor
      tags: [proveedores]
      summary: Obtiene el detalle de un proveedor
      parameters:
        - $ref: "#/components/parameters/ProveedorId"
      responses:
        "200":
          description: Detalle del proveedor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proveedor"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateProveedor
      tags: [proveedores]
      summary: Actualiza un proveedor
      parameters:
        - $ref: "#/components/parameters/ProveedorId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProveedorUpdate"
      responses:
        "200":
          description: Proveedor actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proveedor"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /proveedores/{id}/compras:
    get:
      operationId: listComprasProveedor
      tags: [proveedores]
      summary: Lista compras de un proveedor con estado Pagada/Parcial/Pendiente
      parameters:
        - $ref: "#/components/parameters/ProveedorId"
        - in: query
          name: desde
          schema: { type: string, format: date }
          description: Fecha inicio (inclusive), formato YYYY-MM-DD
        - in: query
          name: hasta
          schema: { type: string, format: date }
          description: Fecha fin (inclusive), formato YYYY-MM-DD
        - in: query
          name: estado
          schema:
            type: string
            enum: [Pagada, Parcial, Pendiente]
          description: Filtrar por estado de pago
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
      responses:
        "200":
          description: Lista paginada de compras del proveedor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProveedorComprasResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /proveedores/{id}/estado-cuenta:
    get:
      operationId: estadoCuentaProveedor
      tags: [proveedores]
      summary: Estado de cuenta cronológico con saldo corrido del proveedor
      parameters:
        - $ref: "#/components/parameters/ProveedorId"
        - in: query
          name: desde
          schema: { type: string, format: date }
          description: Fecha inicio (inclusive), formato YYYY-MM-DD
        - in: query
          name: hasta
          schema: { type: string, format: date }
          description: Fecha fin (inclusive), formato YYYY-MM-DD
      responses:
        "200":
          description: Estado de cuenta con saldo corrido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProveedorEstadoCuenta"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /proveedores/{id}/pagos:
    get:
      operationId: listProveedorPagos
      tags: [proveedores]
      summary: Lista pagos de un proveedor (proveedores_finanzas requerido)
      parameters:
        - $ref: "#/components/parameters/ProveedorId"
        - in: query
          name: desde
          schema: { type: string, format: date }
        - in: query
          name: hasta
          schema: { type: string, format: date }
      responses:
        "200":
          description: Pagos del proveedor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProveedorPagosResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: registrarPagoProveedor
      tags: [proveedores]
      summary: Registra un pago a proveedor (proveedores_finanzas + crear requerido)
      parameters:
        - $ref: "#/components/parameters/ProveedorId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PagoProveedorInput"
      responses:
        "201":
          description: Pago registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagoProveedorRow"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /proveedores/{id}/ajustes:
    post:
      operationId: registrarAjusteProveedor
      tags: [proveedores]
      summary: Registra un ajuste de saldo para un proveedor (ADMIN)
      parameters:
        - $ref: "#/components/parameters/ProveedorId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AjusteProveedorInput"
      responses:
        "201":
          description: Ajuste registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagoProveedorRow"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /proveedores/{id}/estadisticas:
    get:
      operationId: estadisticasProveedor
      tags: [proveedores]
      summary: Estadísticas por periodo (desde y hasta requeridos)
      parameters:
        - $ref: "#/components/parameters/ProveedorId"
        - in: query
          name: desde
          required: true
          schema: { type: string, format: date }
          description: Fecha inicio del periodo, formato YYYY-MM-DD
        - in: query
          name: hasta
          required: true
          schema: { type: string, format: date }
          description: Fecha fin del periodo, formato YYYY-MM-DD
      responses:
        "200":
          description: Estadísticas del periodo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProveedorEstadisticas"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /proveedores/{id}/exportar:
    get:
      operationId: exportarProveedorXlsx
      tags: [proveedores]
      summary: Exporta estado de cuenta del proveedor en XLSX
      parameters:
        - $ref: "#/components/parameters/ProveedorId"
        - in: query
          name: desde
          schema: { type: string, format: date }
          description: Fecha inicio, formato YYYY-MM-DD
        - in: query
          name: hasta
          schema: { type: string, format: date }
          description: Fecha fin, formato YYYY-MM-DD
      responses:
        "200":
          description: Archivo XLSX del estado de cuenta
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    ProveedorId:
      name: id
      in: path
      required: true
      schema:
        type: number
...
  responses:
    Unauthorized:
      description: Sesión no válida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Sin permisos
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ValidationError:
      description: Datos inválidos
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    ProveedoresResumen:
      type: object
      required: [totalProveedores, proveedoresConSaldo, totalDeuda, totalPagado, comprasMes, items]
      properties:
        totalProveedores: { type: number }
        proveedoresConSaldo: { type: number }
        totalDeuda: { type: string }
        totalPagado: { type: string }
        comprasMes: { type: string, description: "Total comprado al conjunto de proveedores en el mes actual" }
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProveedorMetricas"
    ProveedoresListResult:
      type: object
      required: [totalProveedores, items]
      properties:
        totalProveedores: { type: number }
        proveedoresConSaldo: { type: number }
        totalDeuda: { type: string }
        totalPagado: { type: string }
        comprasMes: { type: string, description: "Total comprado al conjunto de proveedores en el mes actual" }
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProveedorMetricas"
...